requirements.txt
numpy~=1.19.1
pillow~=7.2.0
scipy~=1.5.2
logzero~=1.6.3
pathlib2~=2.3.5